Refactor app to split into components
