+++ src/Components/friends-list.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export const FriendsList = ({data}) => {
+    return (
+        <div>
+          <h2>Friends</h2>
+            <ul>
+            {
+              data.map((friend) => <li>{friend.first}</li>)
+            }
+            </ul>
+        </div>
+    )
+};
