--- src/Components/list.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-export const List = ({data}) => {
-  return (
-    <ul>
-      {data.map(item => <li>{item}</li>)}
-    </ul>)
-};
